Move client-related functions to the client namespace from the util
