+++ libdatastore/src/util.rs
@@ -1,5 +1,4 @@
 use crate::resources::DataItemMeta;
-use futures_util::StreamExt;
 use libc::{AT_FDCWD, timespec, utimensat};
 use serde::Serialize;
 use sha2::{Digest, Sha256};
@@ -12,7 +11,6 @@ use std::{
     path::{Path, PathBuf},
     time::{Duration, SystemTime, UNIX_EPOCH},
 };
-use tokio::io::AsyncWriteExt;
 
 /// Returns the current Unix timestamp in seconds.
 pub(crate) fn unix_now() -> u64 {
@@ -129,71 +127,4 @@ pub fn set_file_attrs(meta: &DataItemMeta, dst: impl AsRef<Path>) -> io::Result<
     }
 
     Ok(())
-}
-
-/// Stream-download a blob from the master and write it atomically to `dst`.
-/// - No buffering whole file in memory
-/// - Writes to `dst.tmp` then renames
-///
-/// Example usage:
-/// ```
-/// atomic_download("http://master/store/sha256hash/blob", "/your/bin").await?;
-/// ```
-pub async fn atomic_download(url: &str, dst: impl AsRef<Path>) -> io::Result<()> {
-    let dst = dst.as_ref();
-    let tmp = dst.with_extension("tmp");
-
-    if let Some(parent) = dst.parent() {
-        tokio::fs::create_dir_all(parent).await?;
-    }
-
-    let client = reqwest::Client::new();
-    let resp = client.get(url).send().await.map_err(|e| io::Error::other(e.to_string()))?;
-
-    if !resp.status().is_success() {
-        let status = resp.status();
-        let body = resp.text().await.unwrap_or_default();
-        return Err(io::Error::other(format!("HTTP {status}: {body}")));
-    }
-
-    let mut file = tokio::fs::File::create(&tmp).await?;
-    let mut stream = resp.bytes_stream();
-
-    while let Some(chunk) = stream.next().await {
-        let chunk = chunk.map_err(|e| io::Error::other(e.to_string()))?;
-        file.write_all(&chunk).await?;
-    }
-
-    file.flush().await?;
-    drop(file);
-
-    tokio::fs::rename(&tmp, dst).await?;
-    Ok(())
-}
-
-/// Download a blob by its original filename. This is a convenience function that first resolves the filename to a SHA256 hash, then downloads the blob by hash.
-/// This is not atomic by itself, but relies on the underlying `atomic_download` to ensure atomicity of the file write.
-/// Example usage:
-/// ```
-/// let meta = download_by_name("http://localhost:8080", "myfile.txt").await?;
-/// ```
-pub async fn download_by_name(master: &str, fname: &str) -> io::Result<DataItemMeta> {
-    let client = reqwest::Client::new();
-    let url = format!("{master}/store/resolve?fname={}", urlencoding::encode(fname));
-
-    let meta: DataItemMeta = client
-        .get(url)
-        .send()
-        .await
-        .map_err(|e| io::Error::other(e.to_string()))?
-        .json()
-        .await
-        .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e.to_string()))?;
-
-    let blob_url = format!("{master}/store/{}/blob", meta.sha256);
-
-    atomic_download(&blob_url, fname).await?;
-    set_file_attrs(&meta, fname)?;
-
-    Ok(meta)
 }
